+++ R/mt_aeqd_crs.R
@@ -15,6 +15,7 @@
 #' @examples
 #' mt_aeqd_crs(center = c(10, 45))
 #' mt_aeqd_crs(center = sf::st_point(c(10, 45)), units = "km")
+#' @examplesIf parallel::detectCores() < 9
 #' m <- mt_read(mt_example())
 #' mt_aeqd_crs(center = sf::st_geometry(m)[5])
 #' mt_aeqd_crs(m)