+++ R/mt_read.R
@@ -34,18 +34,16 @@ NULL
 #' @export
 #' @examples
 #' path_fishers <- mt_example()
-#' mt_read(path_fishers, num_threads = 2)
-#' ## Note num_threads is used to reduce the number of threads used.
-#' ## Normally it can be left untouched but here CRAN tests are
-#' ## sped up.
+  #' @examplesIf vroom:::vroom_threads() < 9
+#' mt_read(path_fishers)
 #'
 #' ## Reduce the mount of data read this might provide memory advantages
 #' ## and speed up reading
 #' mt_read(path_fishers, col_select = c(
 #'   "location-long", "location-lat",
-#'   "timestamp", "individual-local-identifier",
-#'   num_threads = 2
+#'   "timestamp", "individual-local-identifier"
 #' ))
+#'
 #' \dontrun{
 #' mt_read("~/Downloads/Galapagos Albatrosses.csv")
 #' ## Exclude the column 'eobs:accelerations-raw'